refactor(filter-button): move filter buttons to src
