+++ packages/components/src/Filter/Filter/docs/Filter.stories.tsx
@@ -2,7 +2,8 @@ import React, { useState } from "react"
 import { ComponentMeta, ComponentStory, Story } from "@storybook/react"
 import isChromatic from "chromatic"
 import { StickerSheet } from "../../../../../../storybook/components/StickerSheet"
-import { Filter, FilterButton, FilterContents } from "../../index"
+import { FilterButton } from "../../../FilterButton"
+import { Filter, FilterContents } from "../../index"
 
 const IS_CHROMATIC = isChromatic()
