+++ packages/components/src/Filter/index.ts
@@ -1,3 +1,2 @@
 export * from "./Filter"
-export * from "./FilterButton"
 export * from "./FilterContents"
