+++ packages/components/src/FilterButton/FilterButtonRemovable/FilterButtonRemovable.tsx
@@ -2,9 +2,9 @@ import React, { forwardRef } from "react"
 import { Icon } from "@kaizen/component-library"
 import iconClear from "@kaizen/component-library/icons/clear.icon.svg"
 import { Tooltip } from "@kaizen/draft-tooltip"
-import { ButtonGroup, ButtonGroupProps } from "../../../ButtonGroup"
-import { DataAttributes } from "../../../types"
-import { isRefObject } from "../../../utils/isRefObject"
+import { ButtonGroup, ButtonGroupProps } from "../../ButtonGroup"
+import { DataAttributes } from "../../types"
+import { isRefObject } from "../../utils/isRefObject"
 import { FilterButtonBase } from "../components/FilterButtonBase"
 import { FilterButton, FilterButtonProps } from ".."
